pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=21.0",
    "django>=4.2.0",
    "requests>=2.31.0",
    "telegram>=0.0.1",
]
